refactor: break up main into modules
