+++ src/parsers/mod.rs
@@ -1,0 +1,2 @@
+pub mod fastx;
+pub mod kraken;
